+++ packages/headless/src/api/commerce/facet-search/facet-search-request.ts
@@ -1,0 +1,6 @@
+import {FacetIdParam, FacetQueryParam} from '../commerce-api-params';
+import {CommerceSearchRequest} from '../search/request';
+
+export type CommerceFacetSearchRequest = CommerceSearchRequest &
+  FacetIdParam &
+  FacetQueryParam;