+++ packages/headless/src/controllers/commerce/core/facets/numeric/headless-commerce-numeric-facet.test.ts
@@ -15,6 +15,7 @@ import {
   MockedCommerceEngine,
 } from '../../../../../test/mock-engine-v2';
 import {commonOptions} from '../../../product-listing/facets/headless-product-listing-facet-options';
+import {NumericRangeRequest} from '../headless-core-commerce-facet';
 import {
   CommerceNumericFacet,
   CommerceNumericFacetOptions,
@@ -40,7 +41,9 @@ describe('CommerceNumericFacet', () => {
     facet = buildCommerceNumericFacet(engine, options);
   }
 
-  function setFacetRequest(config: Partial<CommerceFacetRequest> = {}) {
+  function setFacetRequest(
+    config: Partial<CommerceFacetRequest<NumericRangeRequest>> = {}
+  ) {
     state.commerceFacetSet[facetId] = buildMockCommerceFacetSlice({
       request: buildMockCommerceFacetRequest({facetId, type, ...config}),
     });