+++ packages/headless/src/controllers/commerce/core/facets/regular/headless-commerce-regular-facet.test.ts
@@ -10,10 +10,12 @@ import {buildMockCommerceFacetSlice} from '../../../../../test/mock-commerce-fac
 import {buildMockCommerceRegularFacetValue} from '../../../../../test/mock-commerce-facet-value';
 import {buildMockCommerceState} from '../../../../../test/mock-commerce-state';
 import {
-  buildMockCommerceEngine,
   MockedCommerceEngine,
+  buildMockCommerceEngine,
 } from '../../../../../test/mock-engine-v2';
+import {buildMockFacetSearch} from '../../../../../test/mock-facet-search';
 import {commonOptions} from '../../../product-listing/facets/headless-product-listing-facet-options';
+import {FacetValueRequest} from '../headless-core-commerce-facet';
 import {
   CommerceRegularFacet,
   CommerceRegularFacetOptions,
@@ -24,17 +26,22 @@ jest.mock('../../../../../features/facets/facet-set/facet-set-actions');
 
 describe('CommerceRegularFacet', () => {
   const facetId: string = 'regular_facet_id';
-  let options: CommerceRegularFacetOptions;
-  let state: CommerceAppState;
   let engine: MockedCommerceEngine;
+  let state: CommerceAppState;
+  let options: CommerceRegularFacetOptions;
   let facet: CommerceRegularFacet;
 
-  function initFacet() {
-    engine = buildMockCommerceEngine(state);
+  function initEngine(preloadedState = buildMockCommerceState()) {
+    engine = buildMockCommerceEngine(preloadedState);
+  }
+
+  function initCommerceRegularFacet() {
     facet = buildCommerceRegularFacet(engine, options);
   }
 
-  function setFacetRequest(config: Partial<CommerceFacetRequest> = {}) {
+  function setFacetRequest(
+    config: Partial<CommerceFacetRequest<FacetValueRequest>> = {}
+  ) {
     state.commerceFacetSet[facetId] = buildMockCommerceFacetSlice({
       request: buildMockCommerceFacetRequest({facetId, ...config}),
     });
@@ -43,7 +50,13 @@ describe('CommerceRegularFacet', () => {
     ];
   }
 
+  function setFacetSearch() {
+    state.facetSearchSet[facetId] = buildMockFacetSearch();
+  }
+
   beforeEach(() => {
+    jest.resetAllMocks();
+
     options = {
       facetId,
       ...commonOptions,
@@ -51,39 +64,39 @@ describe('CommerceRegularFacet', () => {
 
     state = buildMockCommerceState();
     setFacetRequest();
+    setFacetSearch();
 
-    initFacet();
+    initEngine(state);
+    initCommerceRegularFacet();
   });
 
-  it('initializes', () => {
-    expect(facet).toBeTruthy();
-  });
+  describe('initialization', () => {
+    it('initializes', () => {
+      expect(facet).toBeTruthy();
+    });
 
-  it('exposes #subscribe method', () => {
-    expect(facet.subscribe).toBeTruthy();
-  });
-
-  describe('#toggleSelect', () => {
-    it('dispatches a #toggleSelectFacetValue', () => {
-      const facetValue = buildMockCommerceRegularFacetValue({value: 'TED'});
-      facet.toggleSelect(facetValue);
-
-      expect(toggleSelectFacetValue).toHaveBeenCalledWith({
-        facetId,
-        selection: facetValue,
-      });
+    it('exposes #subscribe method', () => {
+      expect(facet.subscribe).toBeTruthy();
     });
   });
 
-  describe('#toggleExclude', () => {
-    it('dispatches a #toggleExcludeFacetValue', () => {
-      const facetValue = buildMockCommerceRegularFacetValue({value: 'TED'});
-      facet.toggleExclude(facetValue);
+  it('#toggleSelect dispatches #toggleSelectFacetValue with correct payload', () => {
+    const facetValue = buildMockCommerceRegularFacetValue({value: 'TED'});
+    facet.toggleSelect(facetValue);
 
-      expect(toggleExcludeFacetValue).toHaveBeenCalledWith({
-        facetId,
-        selection: facetValue,
-      });
+    expect(toggleSelectFacetValue).toHaveBeenCalledWith({
+      facetId,
+      selection: facetValue,
+    });
+  });
+
+  it('#toggleExclude dispatches #toggleExcludeFacetValue with correct payload', () => {
+    const facetValue = buildMockCommerceRegularFacetValue({value: 'TED'});
+    facet.toggleExclude(facetValue);
+
+    expect(toggleExcludeFacetValue).toHaveBeenCalledWith({
+      facetId,
+      selection: facetValue,
     });
   });
 });