+++ packages/headless/src/controllers/commerce/search/facets/headless-search-numeric-facet.test.ts
@@ -11,7 +11,10 @@ import {
   buildMockCommerceEngine,
   MockedCommerceEngine,
 } from '../../../../test/mock-engine-v2';
-import {CommerceFacetOptions} from '../../core/facets/headless-core-commerce-facet';
+import {
+  CommerceFacetOptions,
+  NumericRangeRequest,
+} from '../../core/facets/headless-core-commerce-facet';
 import {CommerceNumericFacet} from '../../core/facets/numeric/headless-commerce-numeric-facet';
 import {buildSearchNumericFacet} from './headless-search-numeric-facet';
 
@@ -31,7 +34,9 @@ describe('SearchNumericFacet', () => {
     facet = buildSearchNumericFacet(engine, options);
   }
 
-  function setFacetRequest(config: Partial<CommerceFacetRequest> = {}) {
+  function setFacetRequest(
+    config: Partial<CommerceFacetRequest<NumericRangeRequest>> = {}
+  ) {
     state.commerceFacetSet[facetId] = buildMockCommerceFacetSlice({
       request: buildMockCommerceFacetRequest({facetId, ...config}),
     });