+++ packages/headless/src/controllers/commerce/search/facets/headless-search-regular-facet.test.ts
@@ -8,10 +8,14 @@ import {buildMockCommerceFacetSlice} from '../../../../test/mock-commerce-facet-
 import {buildMockCommerceRegularFacetValue} from '../../../../test/mock-commerce-facet-value';
 import {buildMockCommerceState} from '../../../../test/mock-commerce-state';
 import {
-  buildMockCommerceEngine,
   MockedCommerceEngine,
+  buildMockCommerceEngine,
 } from '../../../../test/mock-engine-v2';
-import {CommerceFacetOptions} from '../../core/facets/headless-core-commerce-facet';
+import {buildMockFacetSearch} from '../../../../test/mock-facet-search';
+import {
+  CommerceFacetOptions,
+  FacetValueRequest,
+} from '../../core/facets/headless-core-commerce-facet';
 import {CommerceRegularFacet} from '../../core/facets/regular/headless-commerce-regular-facet';
 import {buildSearchRegularFacet} from './headless-search-regular-facet';
 
@@ -19,17 +23,22 @@ jest.mock('../../../../features/commerce/search/search-actions');
 
 describe('SearchRegularFacet', () => {
   const facetId: string = 'regular_facet_id';
-  let options: CommerceFacetOptions;
-  let state: CommerceAppState;
   let engine: MockedCommerceEngine;
+  let state: CommerceAppState;
+  let options: CommerceFacetOptions;
   let facet: CommerceRegularFacet;
 
-  function initFacet() {
-    engine = buildMockCommerceEngine(state);
+  function initEngine(preloadedState = buildMockCommerceState()) {
+    engine = buildMockCommerceEngine(preloadedState);
+  }
+
+  function initSearchRegularFacet() {
     facet = buildSearchRegularFacet(engine, options);
   }
 
-  function setFacetRequest(config: Partial<CommerceFacetRequest> = {}) {
+  function setFacetRequest(
+    config: Partial<CommerceFacetRequest<FacetValueRequest>> = {}
+  ) {
     state.commerceFacetSet[facetId] = buildMockCommerceFacetSlice({
       request: buildMockCommerceFacetRequest({facetId, ...config}),
     });
@@ -38,15 +47,23 @@ describe('SearchRegularFacet', () => {
     ];
   }
 
+  function setFacetSearch() {
+    state.facetSearchSet[facetId] = buildMockFacetSearch();
+  }
+
   beforeEach(() => {
+    jest.resetAllMocks();
+
     options = {
       facetId,
     };
 
     state = buildMockCommerceState();
     setFacetRequest();
+    setFacetSearch();
 
-    initFacet();
+    initEngine(state);
+    initSearchRegularFacet();
   });
 
   it('initializes', () => {
@@ -90,6 +107,31 @@ describe('SearchRegularFacet', () => {
   describe('#showLessValues', () => {
     it('dispatches #executeSearch', () => {
       facet.showLessValues();
+
+      expect(executeSearch).toHaveBeenCalled();
+    });
+  });
+
+  describe('#facetSearch', () => {
+    it('#facetSearch.select dispatches #executeSearch', () => {
+      const value = 'ted';
+      facet.facetSearch.select({
+        count: 0,
+        displayValue: value,
+        rawValue: value,
+      });
+
+      expect(executeSearch).toHaveBeenCalled();
+    });
+
+    it('#facetSearch.exclude dispatches #executeSearch', () => {
+      const value = 'ted';
+      facet.facetSearch.exclude({
+        count: 0,
+        displayValue: value,
+        rawValue: value,
+      });
+
       expect(executeSearch).toHaveBeenCalled();
     });
   });
@@ -104,6 +146,15 @@ describe('SearchRegularFacet', () => {
     it('#state.isLoading uses #isFacetLoadingResponseSelector', () => {
       state.commerceSearch.isLoading = true;
       expect(facet.state.isLoading).toBe(true);
+    });
+
+    it('#state.facetSearch exposes the facet search state', () => {
+      expect(facet.state.facetSearch).toBeTruthy();
+      expect(facet.state.facetSearch.isLoading).toBe(false);
+
+      state.facetSearchSet[facetId].isLoading = true;
+      initSearchRegularFacet();
+      expect(facet.state.facetSearch.isLoading).toBe(true);
     });
   });
 });