+++ packages/headless/src/controllers/context/headless-context.ssr.ts
@@ -5,14 +5,17 @@ import {Context, buildContext} from './headless-context';
 
 export * from './headless-context';
 
+export interface ContextDefinition
+  extends ControllerDefinitionWithProps<SearchEngine, Context, ContextProps> {}
+
 /**
- * @internal
- */
-export const defineContext = (): ControllerDefinitionWithProps<
-  SearchEngine,
-  Context,
-  ContextProps
-> => ({
-  buildWithProps: (engine, props) =>
-    buildContext(engine, {initialState: props.initialState}),
-});
+ * Defines a `Context` controller instance.
+ *
+ * @returns The `Context` controller definition.
+ * */
+export function defineContext(): ContextDefinition {
+  return {
+    buildWithProps: (engine, props) =>
+      buildContext(engine, {initialState: props.initialState}),
+  };
+}