+++ packages/headless/src/controllers/results-per-page/headless-results-per-page.ssr.ts
@@ -8,11 +8,19 @@ import {
 
 export * from './headless-results-per-page';
 
+export interface ResultsPerPageDefinition
+  extends ControllerDefinitionWithoutProps<SearchEngine, ResultsPerPage> {}
+
 /**
- * @internal
- */
-export const defineResultsPerPage = (
+ * Defines a `ResultsPerPage` controller instance.
+ *
+ * @param props - The configurable `ResultsPerPage` properties.
+ * @returns The `ResultsPerPage` controller definition.
+ * */
+export function defineResultsPerPage(
   props?: ResultsPerPageProps
-): ControllerDefinitionWithoutProps<SearchEngine, ResultsPerPage> => ({
-  build: (engine) => buildResultsPerPage(engine, props),
-});
+): ResultsPerPageDefinition {
+  return {
+    build: (engine) => buildResultsPerPage(engine, props),
+  };
+}