+++ packages/headless/src/features/commerce/facets/facet-set/facet-set-slice.ts
@@ -9,6 +9,11 @@ import {
   NumericRangeRequest,
 } from '../../../../controllers/commerce/core/facets/headless-core-commerce-facet';
 import {deselectAllBreadcrumbs} from '../../../breadcrumb/breadcrumb-actions';
+import {CategoryFacetValueRequest} from '../../../facets/category-facet-set/interfaces/request';
+import {
+  excludeFacetSearchResult,
+  selectFacetSearchResult,
+} from '../../../facets/facet-search-set/specific/specific-facet-search-actions';
 import {
   toggleExcludeFacetValue,
   toggleSelectFacetValue,
@@ -35,7 +40,10 @@ import {
   CommerceFacetSetState,
   getCommerceFacetSetInitialState,
 } from './facet-set-state';
-import {CommerceFacetRequest} from './interfaces/request';
+import {
+  AnyCommerceFacetRequest,
+  CommerceFacetRequest,
+} from './interfaces/request';
 import {AnyFacetResponse, RegularFacetValue} from './interfaces/response';
 
 export const commerceFacetSetReducer = createReducer(
@@ -49,15 +57,15 @@ export const commerceFacetSetReducer = createReducer(
         const {facetId, selection} = action.payload;
         const facetRequest = state[facetId]?.request;
 
-        if (!facetRequest || facetRequest.type !== 'regular') {
+        if (!facetRequest || !ensureRegularFacetRequest(facetRequest)) {
           return;
         }
 
         facetRequest.preventAutoSelect = true;
 
-        const existingValue = (
-          facetRequest.values as WritableDraft<FacetValueRequest>[]
-        ).find((req) => req.value === selection.value);
+        const existingValue = facetRequest.values.find(
+          (req) => req.value === selection.value
+        );
         if (!existingValue) {
           insertNewValue(facetRequest, selection);
           return;
@@ -69,14 +77,14 @@ export const commerceFacetSetReducer = createReducer(
         const {facetId, selection} = action.payload;
         const facetRequest = state[facetId]?.request;
 
-        if (!facetRequest || facetRequest.type !== 'numericalRange') {
+        if (!facetRequest || !ensureNumericFacetRequest(facetRequest)) {
           return;
         }
 
         facetRequest.preventAutoSelect = true;
 
         const existingValue = findExactRangeValue(
-          facetRequest.values as NumericRangeRequest[],
+          facetRequest.values,
           selection
         );
         if (!existingValue) {
@@ -89,14 +97,14 @@ export const commerceFacetSetReducer = createReducer(
         const {facetId, selection} = action.payload;
         const facetRequest = state[facetId]?.request;
 
-        if (!facetRequest || facetRequest.type !== 'dateRange') {
+        if (!facetRequest || !ensureDateFacetRequest(facetRequest)) {
           return;
         }
 
         facetRequest.preventAutoSelect = true;
 
         const existingValue = findExactRangeValue(
-          facetRequest.values as DateRangeRequest[],
+          facetRequest.values,
           selection
         );
         if (!existingValue) {
@@ -110,15 +118,15 @@ export const commerceFacetSetReducer = createReducer(
         const {facetId, selection} = action.payload;
         const facetRequest = state[facetId]?.request;
 
-        if (!facetRequest || facetRequest.type !== 'regular') {
+        if (!facetRequest || !ensureRegularFacetRequest(facetRequest)) {
           return;
         }
 
         facetRequest.preventAutoSelect = true;
 
-        const existingValue = (
-          facetRequest.values as WritableDraft<FacetValueRequest>[]
-        ).find((req) => req.value === selection.value);
+        const existingValue = facetRequest.values.find(
+          (req) => req.value === selection.value
+        );
         if (!existingValue) {
           insertNewValue(facetRequest, selection);
           return;
@@ -130,14 +138,14 @@ export const commerceFacetSetReducer = createReducer(
         const {facetId, selection} = action.payload;
         const facetRequest = state[facetId]?.request;
 
-        if (!facetRequest || facetRequest.type !== 'numericalRange') {
+        if (!facetRequest || !ensureNumericFacetRequest(facetRequest)) {
           return;
         }
 
         facetRequest.preventAutoSelect = true;
 
         const existingValue = findExactRangeValue(
-          facetRequest.values as NumericRangeRequest[],
+          facetRequest.values,
           selection
         );
         if (!existingValue) {
@@ -151,14 +159,14 @@ export const commerceFacetSetReducer = createReducer(
         const {facetId, selection} = action.payload;
         const facetRequest = state[facetId]?.request;
 
-        if (!facetRequest || facetRequest.type !== 'dateRange') {
+        if (!facetRequest || !ensureDateFacetRequest(facetRequest)) {
           return;
         }
 
         facetRequest.preventAutoSelect = true;
 
         const existingValue = findExactRangeValue(
-          facetRequest.values as DateRangeRequest[],
+          facetRequest.values,
           selection
         );
         if (!existingValue) {
@@ -169,14 +177,73 @@ export const commerceFacetSetReducer = createReducer(
         updateExistingFacetValueState(existingValue, 'exclude');
       })
       // TODO: toggleExcludeCategoryFacetValue
-      .addCase(updateFacetNumberOfValues, (state, action) => {
-        const {facetId, numberOfValues} = action.payload;
+      .addCase(selectFacetSearchResult, (state, action) => {
+        const {facetId, value} = action.payload;
+        const facetRequest = state[facetId]?.request;
 
-        if (!state[facetId]?.request) {
+        if (
+          !facetRequest ||
+          (facetRequest.type !== 'regular' &&
+            facetRequest.type !== 'hierarchical')
+        ) {
           return;
         }
 
-        state[facetId].request.numberOfValues = numberOfValues;
+        const {rawValue} = value;
+
+        facetRequest.preventAutoSelect = true;
+
+        const existingValue = facetRequest.values.find(
+          (v) =>
+            (v as FacetValueRequest | CategoryFacetValueRequest).value ===
+            rawValue
+        );
+
+        if (!existingValue) {
+          insertNewValue(facetRequest, {state: 'selected', value: rawValue});
+          return;
+        }
+
+        updateExistingFacetValueState(existingValue, 'select');
+      })
+      .addCase(excludeFacetSearchResult, (state, action) => {
+        const {facetId, value} = action.payload;
+        const facetRequest = state[facetId]?.request;
+
+        if (
+          !facetRequest ||
+          (facetRequest.type !== 'regular' &&
+            facetRequest.type !== 'hierarchical')
+        ) {
+          return;
+        }
+
+        const {rawValue} = value;
+
+        facetRequest.preventAutoSelect = true;
+
+        const existingValue = facetRequest.values.find(
+          (v) =>
+            (v as FacetValueRequest | CategoryFacetValueRequest).value ===
+            rawValue
+        );
+
+        if (!existingValue) {
+          insertNewValue(facetRequest, {state: 'excluded', value: rawValue});
+          return;
+        }
+
+        updateExistingFacetValueState(existingValue, 'exclude');
+      })
+      .addCase(updateFacetNumberOfValues, (state, action) => {
+        const {facetId, numberOfValues} = action.payload;
+        const facetRequest = state[facetId]?.request;
+
+        if (!facetRequest) {
+          return;
+        }
+
+        facetRequest.numberOfValues = numberOfValues;
       })
       .addCase(updateFacetIsFieldExpanded, (state, action) => {
         const {facetId, isFieldExpanded} = action.payload;
@@ -199,6 +266,24 @@ export const commerceFacetSetReducer = createReducer(
       .addCase(setUser, resetAllFacetValues);
   }
 );
+
+function ensureRegularFacetRequest(
+  facetRequest: AnyCommerceFacetRequest
+): facetRequest is CommerceFacetRequest<FacetValueRequest> {
+  return facetRequest.type === 'regular';
+}
+
+function ensureNumericFacetRequest(
+  facetRequest: AnyCommerceFacetRequest
+): facetRequest is CommerceFacetRequest<NumericRangeRequest> {
+  return facetRequest.type === 'numericalRange';
+}
+
+function ensureDateFacetRequest(
+  facetRequest: AnyCommerceFacetRequest
+): facetRequest is CommerceFacetRequest<DateRangeRequest> {
+  return facetRequest.type === 'dateRange';
+}
 
 function handleQueryFulfilled(
   state: WritableDraft<CommerceFacetSetState>,
@@ -247,7 +332,7 @@ function updateStateFromFacetResponse(
   const facetId = facetResponse.facetId ?? facetResponse.field;
   let facetRequest = state[facetId]?.request;
   if (!facetRequest) {
-    state[facetId] = {request: {} as CommerceFacetRequest};
+    state[facetId] = {request: {} as AnyCommerceFacetRequest};
     facetRequest = state[facetId].request;
     facetRequest.initialNumberOfValues = facetResponse.values.length;
   } else {
@@ -281,7 +366,7 @@ function getFacetRequestValuesFromFacetResponse(
 }
 
 function insertNewValue(
-  facetRequest: CommerceFacetRequest,
+  facetRequest: AnyCommerceFacetRequest,
   facetValue: FacetValueRequest | NumericRangeRequest | DateRangeRequest
 ) {
   const {type, values} = facetRequest;