+++ packages/quantic/force-app/main/default/lwc/quanticResultLink/quanticResultLink.js
@@ -8,6 +8,11 @@ import {ResultUtils} from 'c/quanticUtils';
 import {NavigationMixin} from 'lightning/navigation';
 import {LightningElement, api} from 'lwc';
 
+/**
+ * Some document types cannot be opened directly in Salesforce, but we need to open their parent record, as is the case for the Case Comment document type.
+ */
+const documentTypesRequiringParentRecord = ['CaseComment'];
+
 /** @typedef {import("coveo").Result} Result */
 /** @typedef {import("coveo").SearchEngine} SearchEngine */
 
@@ -139,7 +144,18 @@ export default class QuanticResultLink extends NavigationMixin(
     if (this.result?.raw?.sfkbid && this.result?.raw?.sfkavid) {
       return this.result.raw.sfkavid;
     }
+    if (this.shouldOpenParentRecord) {
+      return this.result?.raw?.sfparentid;
+    }
     return this.result.raw.sfid;
+  }
+
+  get shouldOpenParentRecord() {
+    return (
+      documentTypesRequiringParentRecord.includes(
+        this.result?.raw?.documenttype
+      ) && this.result?.raw?.sfparentid
+    );
   }
 
   /**