+++ packages/quantic/CHANGELOG.md
@@ -1,3 +1,21 @@
+# 2.45.0 (2024-02-21)
+
+### Bug Fixes
+
+- **quantic:** adding support to opening case comments in Salesforce sub tabs ([#3626](https://github.com/coveo/ui-kit/issues/3626)) ([ec0c2dc](https://github.com/coveo/ui-kit/commits/ec0c2dc261017fecb1bfd7c219bf122448ab061f))
+- **Quantic:** Should not return placeholder instead of field value ([#3615](https://github.com/coveo/ui-kit/issues/3615)) ([ec47b16](https://github.com/coveo/ui-kit/commits/ec47b16dc7b2c74fbc2cc38be99f5d1e3f302384))
+
+### Features
+
+- **quantic:** recommendations controller APEX class created ([#3611](https://github.com/coveo/ui-kit/issues/3611)) ([cd03248](https://github.com/coveo/ui-kit/commits/cd03248ba6dec9869be5d9cf440833dcde49a202))
+- **quantic:** styling rules updated to support salesforce community builder customization ([#3603](https://github.com/coveo/ui-kit/issues/3603)) ([25002cd](https://github.com/coveo/ui-kit/commits/25002cda865b3492a6a531a43997bc1426540e8a))
+
+## 2.44.3 (2024-02-15)
+
+### Bug Fixes
+
+- **quantic:** fixed issue with timeframe facet not communicating properly with the refine toggle ([#3588](https://github.com/coveo/ui-kit/issues/3588)) ([4856465](https://github.com/coveo/ui-kit/commits/4856465e87371d871696890c96827570b163fdbb))
+
 ## 2.44.2 (2024-02-14)
 
 ## 2.44.1 (2024-02-09)